cli: isolate state db recovery flow
